refactor: add threadpool_base to include all common code from threadpool implementation
